pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.2",
    "firebase-admin>=6.7.0",
    "google-auth>=2.38.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pycryptodome>=3.22.0",
    "pyrebase>=3.0.18",
    "pyrebase4>=4.8.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "six>=1.17.0",
    "statsmodels>=0.14.4",
    "streamlit>=1.44.0",
]
